+++ CHANGES.rst
@@ -41,6 +41,10 @@ Changes
 - Change to no longer register example *Test*-taxonomy on install
   [datakurre]
 
+- Fix issue where taxonomy indexer adapter was not properly unregistered from
+  the persistent local registry on taxonomy removal
+  [datakurre]
+
 
 1.4.4 (2016-11-29)
 ------------------